refactor(ai-002): split index.js into providerInfo / dispatcher / vision modules
